ci-vuln: partial tree refactoring
