+++ openquakeplatform/openquakeplatform/test/config.py.tmpl
@@ -1,4 +1,4 @@
-pla_basepath="http://localhost:8000"
+pla_basepath="http://oq-platform-mn.gem.lan:8000"
 pla_user="admin"
 pla_passwd="admin"
 # path of debugger .xpi file extension
